MAJOR CHANGE
"@everyone

MAJOR ISSUE

When we were changing our Docker images to our custom one, we were using a outdated version that caused it to not work. I have detected the issue and have tested it on NodeJS v20. Everything else will be slowly updated and we are very sorry for this. If you encounter any more issues please contact me in a ticket or ⁠❓・community-support . Thank you." - Lezetho

https://discord.com/channels/852220533499494411/1073640508083159132/1265020157227434126 - sent in https://discord.embotic.xyz